Refactor AbstractHTML/JSON and AlbumRipper to have a common ancestor
